reformatted code and moved classes to their appropriate package
