--- src/main/java/application/auth/MyCorsConfiguration.java
@@ -1,37 +1,0 @@
-package application.auth;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.Ordered;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-import org.springframework.web.filter.CorsFilter;
-
-@Configuration
-public class MyCorsConfiguration  {
-    private static Logger logger =  LoggerFactory.getLogger(MyCorsConfiguration.class);
-
-	@Bean
-    public FilterRegistrationBean<CorsFilter> corsFilter() {
-		logger.info("corsFilter()");
-		
-		final UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-		final CorsConfiguration config = new CorsConfiguration();
-		
-		config.setAllowCredentials(true);
-		config.addAllowedOrigin("*");
-		config.addAllowedHeader("*");
-		config.addAllowedMethod("OPTIONS");
-		config.addAllowedMethod("POST");
-		config.addAllowedMethod("GET");
-		source.registerCorsConfiguration("/**", config);
-		
-		final FilterRegistrationBean<CorsFilter> bean = new FilterRegistrationBean<CorsFilter>(new CorsFilter(source));
-		bean.setOrder(Ordered.HIGHEST_PRECEDENCE);
-		
-		return bean;
-	}
-}
