+++ src/main/java/application/auth/config/SwaggerConfiguration.java
@@ -1,8 +1,7 @@
-package application;
+package application.auth.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-
 import springfox.documentation.builders.ApiInfoBuilder;
 import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
@@ -14,8 +13,8 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2;
 @Configuration
 @EnableSwagger2
 public class SwaggerConfiguration {
-	
-	@Bean
+
+    @Bean
     public Docket api() {
 
         return new Docket(DocumentationType.SWAGGER_2).select()
