+++ src/main/java/application/auth/services/AboutServiceImpl.java
@@ -1,15 +1,14 @@
 package application.auth.services;
 
+import application.auth.models.About;
 import org.springframework.stereotype.Service;
 
-import application.auth.models.About;
-
 @Service
-public class AboutServiceImpl implements AboutService{
+public class AboutServiceImpl implements AboutService {
 
-	@Override
-	public About getInfo() {
-		return new About("Auth Service", "Storefront", "Authorization and Authentication");
-	}
+    @Override
+    public About getInfo() {
+        return new About("Auth Service", "Storefront", "Authorization and Authentication");
+    }
 
 }
